feat: deterministic game sessions via --seed with per-VM seeded Lua and RNG
- add GameSessionConfig with 64-bit session seed passed into Room and RoomRuntime
- support --seed <n> in client and headless server entry points, seeding global RNG, QHash, and per-session seeds
- derive per-session seeds from a base (seed + session index); log #GameSeed on room start
- add lua_newstate_seeded with deterministic hash seed and seeded CreateLuaState overloads
- route Lua math.random/randomseed through the game RNG so gameplay VM randomness is reproducible
- widen GameRng/qsanSeedRandom/room-runtime/ai-runtime seed APIs to quint64
- drop time()-based re-seeding in boss-mode, godLottery, and addRobotCommand

+++ src/core/game-rng.h
@@ -8,7 +8,7 @@ class GameRng
 public:
     GameRng();
 
-    void seed(quint32 seed);
+    void seed(quint64 seed);
     int bounded(int upperExclusive);
 
     static GameRng *current();
@@ -32,6 +32,6 @@ private:
 };
 
 int qsanRandomBounded(int upperExclusive);
-void qsanSeedRandom(quint32 seed);
+void qsanSeedRandom(quint64 seed);
 
 #endif
